Add files via upload
I was deeply saddened to learn that Python didn't have switch case so I put one together that is
pleasant on the eyes and it works and it's simple. Einstein said simplicity is beauty. 
If you add a function call inside each of the cases you can have pages of code run in each case if you wish.